# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is the Transformers back-end server for purposes of the LimeSpot front-end assignment.",
      "version": "1.0.0",
      "title": "Teletran-1 API",
      "termsOfService": "http://transformers.morethanmeetsthe.io/",
      "contact": {
        "email": "optimus@theark.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host": "localhost:3456",
    "basePath": "",
...
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/transformers/{id}": {
        "get": {
          "tags": [
            "transformer"
          ],
          "summary": "Gets a Transformer by ID",
          "description": "Gets a Transformer corresponding to the ID passed in.",
          "operationId": "getTransformerByID",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "ID of the Transformer in question",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Transformer"
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "transformer"
          ],
          "summary": "Update an existing transformer",
          "description": "Ultra-cool note: the example PUT provided below adjusts Scorponok's stats to the original printed Tech Spec and not the corrected version implied by the readout from the Headmaster display on the chest of the toy.",
          "operationId": "updateTransformer",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Existing Transformer object for editing",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Transformer"
              }
            },
            {
              "in": "path",
              "name": "id",
              "description": "ID of the Transformer in question",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Transformer not found"
            },
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Transformer"
              }
            }
          }
        },
        "delete": {
          "tags": [
            "transformer"
          ],
          "summary": "Delete an existing transformer",
          "description": "This deletes a Transformer from the database.",
          "operationId": "deleteTransformer",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "ID of the Transformer in question",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Transformer not found"
            },
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/Transformer"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
        "Transformer" : {
          "properties": {
              "allegiance": {
                  "type": "string",
                  "description": "Allegiance",
                  "enum": ["Autobot","Decepticon"],
                  "example": "Decepticon"
              },
              "id": {
                "type": "integer",
                "minimum": 1,
                "example": 9
              },
              "name": {
                  "type": "string",
                  "description": "Name",
                  "example": "Scorponok"
              },
              "strength": {
                  "type": "integer",
                  "description": "Strength",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 8
              },
              "intelligence": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 7
              },
              "speed": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 2
              },
              "endurance": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 8
              },
              "rank": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 8
              },
              "courage": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 8
              },
              "firepower": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 7
              },
              "skill": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 10,
                  "example": 7
              }
          }
      }
    },
...
  }